requirements.txt
python-i18n==0.3.9
Babel>=2.9.0
